Moved all DI components to 1 AppModule.
Only the view models should not be singleton.
