+++ android/app/src/main/java/com/koleff/kare_android/common/di/AppModule.kt
@@ -5,10 +5,14 @@ import com.jakewharton.retrofit2.adapter.kotlin.coroutines.CoroutineCallAdapterF
 import com.koleff.kare_android.common.Constants
 import com.koleff.kare_android.data.datasource.ExerciseDataSource
 import com.koleff.kare_android.data.datasource.ExerciseRemoteDataSource
+import com.koleff.kare_android.data.datasource.WorkoutDataSource
+import com.koleff.kare_android.data.datasource.WorkoutRemoteDataSource
 import com.koleff.kare_android.data.remote.ExerciseApi
 import com.koleff.kare_android.data.remote.WorkoutApi
 import com.koleff.kare_android.data.repository.ExerciseRepositoryImpl
+import com.koleff.kare_android.data.repository.WorkoutRepositoryImpl
 import com.koleff.kare_android.domain.repository.ExerciseRepository
+import com.koleff.kare_android.domain.repository.WorkoutRepository
 import com.koleff.kare_android.ui.view_model.ExerciseViewModel
 import com.squareup.moshi.Moshi
 import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
@@ -91,4 +95,40 @@ object AppModule {
             .build()
             .create(WorkoutApi::class.java)
     }
+
+    @Provides
+    @Singleton
+    fun provideExerciseDataSource(exerciseApi: ExerciseApi): ExerciseDataSource {
+        return ExerciseRemoteDataSource(exerciseApi) //Can swap for local data source...
+    }
+
+    @Provides
+    @Singleton
+    //    @ViewModelScoped
+    fun provideExerciseRepository(exerciseDataSource: ExerciseDataSource): ExerciseRepository {
+        return ExerciseRepositoryImpl(exerciseDataSource)
+    }
+
+    @Provides
+    fun providesExerciseViewModel(exerciseRepository: ExerciseRepository): ExerciseViewModel {
+        return ExerciseViewModel(exerciseRepository)
+    }
+
+    @Provides
+    @Singleton
+    fun provideWorkoutDataSource(workoutApi: WorkoutApi): WorkoutDataSource {
+        return WorkoutRemoteDataSource(workoutApi) //Can swap for local data source...
+    }
+
+    @Provides
+    @Singleton
+    //    @ViewModelScoped
+    fun provideWorkoutRepository(workoutDataSource: WorkoutDataSource): WorkoutRepository {
+        return WorkoutRepositoryImpl(workoutDataSource)
+    }
+
+//    @Provides
+//    fun providesWorkoutViewModel(workoutRepository: WorkoutRepository): WorkoutViewModel {
+//        return WorkoutViewModel(workoutRepository)
+//    }
 }
